requirements.txt
python-multipart
fastapi
uvicorn
openai
python-dotenv
jinja2
python-multipart
openai==0.28